fix unit test and refactor on structure
