+++ app/src/main/java/app/igormatos/botaprarodar/presentation/fullscreenimage/FullscreenImageActivity.kt
@@ -5,7 +5,7 @@ import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import app.igormatos.botaprarodar.R
-import app.igormatos.botaprarodar.common.util.loadPath
+import com.brunotmgomes.ui.extensions.loadPath
 import kotlinx.android.synthetic.main.activity_fullscreen_image.*
