+++ toolkit/src/main/java/com/brunotmgomes/ui/SimpleResult.kt
@@ -1,4 +1,4 @@
-package utils
+package com.brunotmgomes.ui
 
 sealed class SimpleResult<out T> {
     class Success<out T>(val data: T) : SimpleResult<T>()
