+++ toolkit/src/main/java/com/brunotmgomes/ui/extensions/CharSequenceExtension.kt
@@ -1,4 +1,4 @@
-package app.igormatos.botaprarodar.common.util
+package com.brunotmgomes.ui.extensions
 
 import android.util.Patterns
