+++ toolkit/src/main/java/com/brunotmgomes/ui/extensions/LiveDataExtensions.kt
@@ -1,4 +1,4 @@
-package com.brunotmgomes.ui
+package com.brunotmgomes.ui.extensions
 
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.LiveData
