+++ toolkit/src/main/java/com/brunotmgomes/ui/extensions/ShowLoading.kt
@@ -1,4 +1,4 @@
-package utils
+package com.brunotmgomes.ui.extensions
 
 import androidx.appcompat.app.AlertDialog
 import android.content.Context
